+++ vodka/resources/blank_app/application.py
@@ -1,0 +1,5 @@
+import vodka.app
+
+#@vodka.app.register('my_app')
+#class MyApplication(vodka.app.Application):
+#    pass