+++ vodka/resources/blank_app/plugins/example.py
@@ -1,0 +1,6 @@
+import vodka
+import vodka.plugins
+
+#vodka.plugin.register('my_plugin')
+#class MyPlugin(vodka.plugins.PluginBase):
+#    pass